reverse-iterate over IPs for each header
Reverse-iterating over the IPs that accumulate in, for example, the X-Forwarded-For
header, thwarts attackers who try to spoof an IP address by submitting a header with
a bogus IP in the original HTTP request.

For example, if Mallory issues an HTTP request against a server with the
X-Forwarded-For header value set to "8.8.8.8," and two proxy servers append two other
IP addresses to the header value, then when the request finally reaches Django, its
X-Forwarded-For header might look something like this:

    X-Forwarded-For: 8.8.8.8,{REAL CLIENT IP},{PROXY 1 IP}

Whether we reverse-iterate or we forward-iterate over the above list, the goal is to
return an IP address that belongs to a machine other than a server, namely so that we
can derive the IP address of the client issuing the request. However, reverse-iterating
ensures that we return an IP address inserted into the above list by a trusted machine
(i.e., REAL CLIENT IP, inserted into the list by PROXY 1), whereas forward-iterating
would gobble up the first potentially bogus IP it encounters (i.e., 8.8.8.8, inserted
by the user-agent issuing the original HTTP request).