ipware: Adjust Forwarded handling to match RFC 7239
RFC 7239 defines the 'Forwarded' header [1], which is not just a
comma-separated list of IP addresses (which is how ipware has treated
the header since it started watching it in e30af18f, add to precedence
list, 2013-12-23).  This commit adds support for the RFC form using
email.utils.unquote [2,3] and an HTTP-list parser.  I'd intially used
the undocumented parse_http_list [4], but it incorrectly strips
backslashes [5].

The example values I've used in the tests are mostly from (or
excerpted from) the RFC [1].

[1]: https://tools.ietf.org/html/rfc7239
[2]: https://docs.python.org/2/library/email.util.html#email.utils.unquote
[3]: https://docs.python.org/3/library/email.util.html#email.utils.unquote
[4]: https://bugs.python.org/issue33003
[5]: https://bugs.python.org/issue33008